fix: fixing error about casting in height

+++ phase-1-rtiow/src/main.rs
@@ -17,7 +17,7 @@ fn main() {
     // setup canvas
     let aspect_ratio = 16.0 / 9.0;
     let image_width = 400;
-    let image_height = image_width / aspect_ratio as i32;
+    let image_height = (image_width as f64 / aspect_ratio) as i32;
 
     // Camera
     let focal_length = 1.0;
